requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncio==3.4.3
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
mido==1.3.2
openai==1.51.0
packaging==23.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
RPi.GPIO==0.7.1
rpimotorlib==3.2
sniffio==1.3.1
sshkeyboard==2.3.1
tqdm==4.66.5
typing_extensions==4.12.2
